backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic==1.13.1
pydantic
pydantic-settings
python-dotenv
redis==5.0.1
playwright==1.41.0
google-genai
fake-useragent
celery
beautifulsoup4
APScheduler
